ls-files: add %(objecttype) atom to format option
Sometimes users may want to align the feature of
`git ls-files --format` with that of `git ls-tree --format`,
but the %(objecttype) atom is missing in the format option
of git ls-files compared to git ls-tree.

Therefore, the %(objecttype) atom is added to the format option
of git ls-files, which can be used to obtain the object type
of the file which is recorded in the index.

Signed-off-by: ZheNing Hu <adlternative@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>