ls-files: add %(objectsize) atom to format option
Sometimes users may want to align the feature of
`git ls-files --format` with that of `git ls-tree --format`,
but the %(objectsize) and %(objectsize:padded) are missing
in the format option of git ls-files compared to git ls-tree.

Therefore, the %(objecttsize) atom is added to the format
option of git ls-files, which can be used to obtain the
object size of the file which is recorded in the index.
("-" if the object is a `commit` or `tree`) It also
supports a padded format of size with %(objectsize:padded).

Signed-off-by: ZheNing Hu <adlternative@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>